ARGO-583 Adds support for deployment via docker
Up to know we had only one deployment model for the ARGO Messaging
Service via the use of RPMs for CENTOS and RHEL environments. This
commit adds support for orchestrating the AMS deployment via docker.

It provides a way to deploy AMS on any Linux distribution that supports
a recent version of the docker-engine and docker-compose. It has been
tested on Ubuntu 16.04.1.

Dockerized features:

- Run unit tests
- Build linux static binary
- Get certificates via letsencrypt
- Deploy Zookeper, Kafka, MongoDB and AMS